refactor navbar
